<FIX> Memory issue when receiving JoinCFM
1. Fix the linked list part where the memory access was wrong (now still commented out since unused)

2. Add a typecast for the time as it can help with the timer rollover.

+++ MessageProcessor.cpp
@@ -163,7 +163,7 @@ GenericMessage* receiveMessage(DeviceDriver* driver, unsigned long timeout)
     char msgType = -1;
     GenericMessage* msg = nullptr;
 
-    while(getTimeMillis() - startTime < timeout)
+    while((unsigned long)(getTimeMillis() - startTime) < timeout)
     {
         // get first char, check msg type
         msgType = driver->recv();
